refactor: Auth 관련 컴포넌트 중복 styled 코드 리팩토링
